Move sign error tests into Signer-test
